[#563] Address T2b review: fix cooldown, widen quotient precision, extract helper
- Cooldown now gates ALL refreshes in /api/user/onboard regardless of
  forceRefresh flag
- Widen quotient_score from DECIMAL(5,4) to DECIMAL(10,4) to handle
  any score scale
- Extract shared buildUserData() helper to lib/user-data.ts to reduce
  duplication between register-by-wallet and onboard routes
